requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via vehicle-accounting
aiosignal==1.3.2
    # via aiohttp
anyio==4.9.0
    # via httpx
asgiref==3.8.1
    # via
    #   channels
    #   daphne
    #   django
attrs==25.3.0
    # via
    #   aiohttp
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==25.4.16
    # via twisted
branca==0.8.1
    # via folium
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
    # via
    #   cryptography
    #   gevent
channels==4.2.2
    # via vehicle-accounting
charset-normalizer==3.4.2
    # via requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via pytest
constantly==23.10.4
    # via twisted
coverage==7.9.2
    # via pytest-cov
cryptography==45.0.4
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.2.0
    # via vehicle-accounting
diff-match-patch==20241021
    # via django-import-export
django==5.2.1
    # via
    #   channels
    #   django-bootstrap5
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-import-export
    #   django-model-utils
    #   djangorestframework
    #   djangorestframework-gis
    #   djangorestframework-simplejwt
    #   vehicle-accounting
django-admin-rangefilter==0.13.2
    # via vehicle-accounting
django-bootstrap5==25.1
    # via vehicle-accounting
django-debug-toolbar==5.2.0
    # via vehicle-accounting
django-extensions==4.1
    # via vehicle-accounting
django-filter==24.3
    # via djangorestframework-gis
django-import-export==4.3.7
    # via vehicle-accounting
django-model-utils==5.0.0
    # via vehicle-accounting
djangorestframework @ git+https://github.com/encode/django-rest-framework@30384947053b1f2b2c9e82cafd1da934d3442a61
    # via
    #   djangorestframework-gis
    #   djangorestframework-simplejwt
    #   vehicle-accounting
djangorestframework-gis==1.2.0
    # via vehicle-accounting
djangorestframework-simplejwt==5.5.0
    # via vehicle-accounting
execnet==2.1.1
    # via pytest-xdist
factory-boy==3.3.3
    # via
    #   pytest-factoryboy
    #   vehicle-accounting
faker==37.1.0
    # via
    #   factory-boy
    #   vehicle-accounting
folium==0.19.5
    # via vehicle-accounting
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
gevent==25.5.1
    # via vehicle-accounting
gpxpy==1.6.2
    # via vehicle-accounting
greenlet==3.2.3 ; platform_python_implementation == 'CPython'
    # via gevent
gunicorn==23.0.0
    # via vehicle-accounting
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via python-telegram-bot
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
    #   twisted
    #   yarl
incremental==24.7.2
    # via twisted
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via
    #   branca
    #   folium
kafka-python==2.2.15
    # via vehicle-accounting
markupsafe==3.0.2
    # via jinja2
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
numpy==2.2.5
    # via folium
packaging==25.0
    # via
    #   gunicorn
    #   pytest
    #   pytest-factoryboy
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psycopg==3.2.9
    # via vehicle-accounting
psycopg-pool==3.2.6
    # via psycopg
psycopg2-binary==2.9.10
    # via vehicle-accounting
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.22 ; platform_python_implementation != 'PyPy'
    # via cffi
pygments==2.19.2
    # via pytest
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pymemcache==4.0.0
    # via vehicle-accounting
pyopenssl==25.1.0
    # via twisted
pytest==8.4.1
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
    #   pytest-mock
    #   pytest-xdist
    #   vehicle-accounting
pytest-cov==6.2.1
    # via vehicle-accounting
pytest-django==4.11.1
    # via vehicle-accounting
pytest-factoryboy==2.8.1
    # via vehicle-accounting
pytest-mock==3.14.1
    # via vehicle-accounting
pytest-xdist==3.8.0
    # via vehicle-accounting
python-dotenv==1.1.0
    # via vehicle-accounting
python-telegram-bot==22.1
    # via vehicle-accounting
pytz==2025.2
    # via vehicle-accounting
requests==2.32.3
    # via
    #   folium
    #   vehicle-accounting
rx==3.2.0
    # via vehicle-accounting
service-identity==24.2.0
    # via twisted
setuptools==80.9.0
    # via
    #   autobahn
    #   incremental
    #   zope-event
    #   zope-interface
sniffio==1.3.1
    # via anyio
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
tablib==3.8.0
    # via django-import-export
tomli==2.2.1 ; python_full_version <= '3.11'
    # via coverage
twisted==25.5.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.13.2
    # via
    #   anyio
    #   psycopg
    #   psycopg-pool
    #   pyopenssl
    #   pytest-factoryboy
    #   twisted
tzdata==2025.2
    # via
    #   django
    #   faker
    #   psycopg
urllib3==2.4.0
    # via requests
uwsgi==2.0.29
    # via vehicle-accounting
xyzservices==2025.4.0
    # via folium
yarl==1.20.0
    # via aiohttp
zope-event==5.1
    # via gevent
zope-interface==7.2
    # via
    #   gevent
    #   twisted
